event_collector/requirements.txt
flask
PyYaml
celery
requests
redis